requirements.txt
Django==1.8.5
django-bananas==1.0.8
django-enumfields==0.7.3
django-factory-boy==0.1.6
django-model-utils==2.2
django-money==0.7.0
djangorestframework==3.2.0
django-rest-swagger==0.3.4
django-bitfield==1.8.0
enum34==1.0.4
factory-boy==2.5.2
Pillow==2.9.0
pluggy==0.3.0
py==1.4.30
py-moneyed==0.5.0
pyflakes==0.8.1
six==1.9.0
tox==2.1.1
requests==2.8.1

# production & dev
psycopg2==2.5.4
raven==5.7.2
